--- backend/src/main/java/com/imsnacks/Nyeoreumnagi/member/entity/Farm.java
@@ -1,45 +1,0 @@
-package com.imsnacks.Nyeoreumnagi.member.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Getter
-public class Farm {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    @Column(nullable = false)
-    private String state;
-
-    @Column(nullable = false)
-    private String city;
-
-    @Column(nullable = false)
-    private String town;
-
-    @Column(nullable = false)
-    private String address;
-
-    @Column(nullable = false)
-    private Double latitude;
-
-    @Column(nullable = false)
-    private Double longitude;
-
-    @Column(nullable = false)
-    private Integer nx;
-
-    @Column(nullable = false)
-    private Integer ny;
-
-    private String midTempRegionCode;
-
-    @OneToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    Member member;
-}
