+++ backend/src/main/java/com/imsnacks/Nyeoreumnagi/member/repository/FarmRepository.java
@@ -1,6 +1,6 @@
 package com.imsnacks.Nyeoreumnagi.member.repository;
 
-import com.imsnacks.Nyeoreumnagi.member.entity.Farm;
+import com.imsnacks.Nyeoreumnagi.farm.entity.Farm;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
